fix compile fail on riscv64

+++ src/file.c
@@ -42,7 +42,7 @@
 static const char *file_mktag(struct stat *s, char *buf, int len)
 {
     snprintf(buf, len, "\"%" PRIx64 "-%" PRIx64 "-%" PRIx64 "\"",
-             s->st_ino, s->st_size, (uint64_t)s->st_mtime);
+             (uint64_t)s->st_ino, s->st_size, (uint64_t)s->st_mtime);
 
     return buf;
 }
